fix(dns): prevent comment loss when deleting duplicate records

+++ lib/Application/Controller/DeleteRecordController.php
@@ -89,15 +89,19 @@ class DeleteRecordController extends BaseController
                 $dnsRecord = new DnsRecord($this->db, $this->getConfig());
                 $dnsRecord->update_soa_serial($zid);
 
-                $this->recordCommentService->deleteComment($domain_id, $record_info['name'], $record_info['type']);
-
                 if ($this->config('pdnssec_use')) {
                     $zone_name = $dnsRecord->get_domain_name_by_id($zid);
                     $dnssecProvider = DnssecProviderFactory::create($this->db, $this->getConfig());
                     $dnssecProvider->rectifyZone($zone_name);
                 }
 
-                $this->setMessage('edit', 'success', _('The record has been deleted successfully.'));
+                if (!$dnsRecord->has_similar_records($domain_id, $record_info['name'], $record_info['type'], $record_id)) {
+                    $this->recordCommentService->deleteComment($domain_id, $record_info['name'], $record_info['type']);
+                    $this->setMessage('edit', 'success', _('The record has been deleted successfully.'));
+                } else {
+                    $this->setMessage('edit', 'warn', _('The record was deleted but the comment was preserved because similar records exist.'));
+                }
+
                 $this->redirect('index.php', ['page'=> 'edit', 'id' => $zid]);
             }
         }
